files path changed and renamed
