+++ client/browser-app.js
